requirements.txt
Adafruit-bitfield==1.5.1
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-pixelbuf==2.0.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-seesaw==1.16.5
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
Adafruit_GPIO==1.0.3
Adafruit_seesaw==1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-iot-device==2.14.0
binho-host-adapter==0.1.6
blinker==1.9.0
board==1.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==3.0.1
flatbuffers==25.2.10
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
openai==0.28
openmeteo_requests==1.4.0
openmeteo_sdk==1.19.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
pyftdi==0.56.0
pymongo==4.12.0
pyparsing==3.2.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
requests==2.32.3
requests-cache==1.2.1
requests-unixsocket2==1.0.0
retry-requests==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.4.0
Werkzeug==3.1.3
azure-monitor-query==1.4.1
azure-identity==1.21.0
azure-mgmt-monitor==6.0.2